+++ Editor/RigSetup/RigLoaderSceneSetup.cs.meta
