+++ Runtime/RigSetup/InteractionRigSetup.cs
@@ -5,7 +5,7 @@ using Innoactive.Creator.Core.Properties;
 using Innoactive.Creator.Core.Utils;
 using UnityEngine;
 
-namespace Innoactive.Creator.BasicInteraction.Setup
+namespace Innoactive.Creator.BasicInteraction.RigSetup
 {
     /// <summary>
     /// Will setup the interaction rig on awake. Priority is from top to bottom of the list, only rigs which full fill
